Expose utilities in NPM package
