+++ .github/workflows/publish.yml
@@ -49,9 +49,10 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-node-
       - name: Remove dark theme images from README
-        uses: mondeja/strip-gh-theme-links@v3
+        uses: mondeja/strip-gh-theme-links@v4
         with:
           files: README.md
+          keep: light
           strict: true
       - name: Install dependencies
         run: npm i
@@ -75,9 +76,10 @@ jobs:
           export PACKAGE_VERSION=$(cat package.json | grep 'version' | sed 's/[ \",:]//g' | sed 's/version//')
           echo "::set-output name=version::$PACKAGE_VERSION"
       - name: Remove dark theme images from README
-        uses: mondeja/strip-gh-theme-links@v3
+        uses: mondeja/strip-gh-theme-links@v4
         with:
           files: README.md
+          keep: light
           strict: true
       - name: Configure GIT credentials
         run: |
