+++ .npmignore
@@ -12,3 +12,4 @@
 !index.mjs
 !index.d.ts
 !types.d.ts
+!utils.mjs
