+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "simple-icons",
-  "version": "7.20.0",
+  "version": "7.21.0",
   "description": "SVG icons for popular brands https://simpleicons.org",
   "homepage": "https://simpleicons.org",
   "keywords": [
@@ -20,7 +20,8 @@
     },
     "./icons/*": [
       "./icons/*"
-    ]
+    ],
+    "./utils": "./utils.mjs"
   },
   "sideEffects": false,
   "repository": {
