+++ scripts/lint/ourlint.js
@@ -6,7 +6,11 @@
  */
 
 import fakeDiff from 'fake-diff';
-import { getIconsDataString, normalizeNewlines, collator } from '../utils.js';
+import {
+  getIconsDataString,
+  normalizeNewlines,
+  collator,
+} from '../../utils.mjs';
 
 /**
  * Contains our tests so they can be isolated from each other.
