+++ scripts/release/update-cdn-urls.js
@@ -7,7 +7,7 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
-import { getDirnameFromImportMeta } from '../utils.js';
+import { getDirnameFromImportMeta } from '../../utils.mjs';
 
 const __dirname = getDirnameFromImportMeta(import.meta.url);
